refactor(vault): reorganize the god-mode panel into surface tabs
