+++ services/vault/src/dev/GodModePanel.tsx
@@ -1,682 +1,97 @@
 /**
  * GodModePanel (dev / QA only — gated behind NEXT_PUBLIC_FF_GOD_MODE_PANEL).
  *
- * A draggable, floating "god mode" admin panel for exercising UI states during
- * development. It can also pop out into a separate browser window (rendered via
- * a React portal, so it stays in the same React tree and shares state with the
- * app — no cross-window plumbing).
+ * A draggable, floating "god mode" admin panel for exercising UI states
+ * during development. It can also pop out into a separate browser window
+ * (rendered via a React portal, so it stays in the same React tree and
+ * shares state with the app — no cross-window plumbing).
  *
- * It is a controller only: it never renders the cards itself. It manages a list
- * of mock items (each a deposit / collateral / loan at some state) that render
- * inside the REAL dashboard, vaults and loans sections (see demoDeposit.ts).
- * It is mounted once for those routes (see dev/GodModeMount).
+ * The shell only owns chrome: drag, pop-out, open/close, the tab rail, and
+ * the pinned active-overrides summary chip. Every tab's actual controls live
+ * in `./panels/*` and are wired in by `./registry.ts` — see `GodModeMount`.
  *
- * Chrome is intentionally theme-independent (fixed zinc colors) and inline, not
- * routed through copy.ts — none of it is shown to depositors.
+ * Every VISIBLE tab is mounted at once (hidden via the `hidden` attribute,
+ * not conditional rendering) rather than only the active one. Before this
+ * refactor every section rendered simultaneously inside one flat body, so
+ * each dev-store→override-store publish effect ran whenever the panel was
+ * open, regardless of which `<details>` a QA engineer had expanded. Mounting
+ * only the active tab would replicate that bug at the tab level: switching
+ * away from Position would tear down its publish effects (including the
+ * cascade simulator's deliberate unmount-clear) mid-session. Mounting every
+ * tab and hiding the rest keeps that exact "always mounted while the panel is
+ * open" contract; tab switching is pure display.
+ *
+ * Chrome is intentionally theme-independent (fixed zinc colors) and inline,
+ * not routed through copy.ts — none of it is shown to depositors.
  */
 
-import { useTheme } from "next-themes";
-import {
-  type ReactNode,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { createPortal } from "react-dom";
 
-import type { ProtocolStatus } from "@/components/shared/protocolStatus";
-import { setActivityOverride } from "@/overrides/activity";
-import { setArtifactDownloadOverride } from "@/overrides/artifactDownload";
-import {
-  setBorrowCapacityOverride,
-  setHealthFactorOverride,
-} from "@/overrides/borrowCapacity";
-import { setCollateralOverride } from "@/overrides/collateral";
-import { setDepositOverride } from "@/overrides/deposits";
-import { setLoanOverride } from "@/overrides/loans";
-import {
-  setMaxVaultsOverride,
-  setProtocolStatusOverride,
-} from "@/overrides/protocolStatus";
-import { clearArtifactDownloadReceipts } from "@/utils/artifactDownloadStorage";
+import { PANEL_BUTTON_CLASS } from "./panelChrome";
+import { CascadeOverridePublisher } from "./panels/CascadeSimulator";
+import { SegmentButton } from "./panels/segmentButton";
+import type { DevPanelTab } from "./registry";
+import { SummaryChip } from "./SummaryChip";
 
-import {
-  DEBUG_FORCED_MAX_VAULTS,
-  DEBUG_HEALTH_FACTORS,
-  type DebugBorrowCapacityState,
-  setDebugBorrowCapacityStateOverride,
-  setDebugHealthFactorOverride,
-  setDebugMaxVaultsOverride,
-  setDebugProtocolStatusOverride,
-  useDebugBorrowCapacity,
-  useDebugBorrowCapacityStateOverride,
-  useDebugHealthFactorOverride,
-  useDebugMaxVaultsOverride,
-  useDebugProtocolStatusOverride,
-} from "./debugPositionStore";
-import {
-  DEMO_ARTIFACT_SCENARIO_LABELS,
-  DEMO_ARTIFACT_SCENARIOS,
-  type DemoArtifactScenario,
-  demoFetchAndDownloadArtifacts,
-  setArtifactDownloadMockEnabled,
-  setArtifactDownloadScenario,
-  useArtifactDownloadMockEnabled,
-  useArtifactDownloadScenario,
-} from "./demoArtifactDownload";
-import {
-  addDemoItem,
-  amountUnitFor,
-  buildActivitiesDemo,
-  buildCollateralsDemo,
-  buildDepositsDemo,
-  buildLoansDemo,
-  DEMO_BORROW_SYMBOL_OPTIONS,
-  type DemoBorrowSymbol,
-  type DemoCta,
-  type DemoItem,
-  type DemoType,
-  DEPOSIT_ACTIVATION_WINDOW_SCENARIO_COUNT,
-  DEPOSIT_DIFFERENT_WALLET_SCENARIO_COUNT,
-  DEPOSIT_EXPIRED_SCENARIO_COUNT,
-  DEPOSIT_FLOW_SCENARIO_COUNT,
-  itemSectionHint,
-  removeDemoItem,
-  scenariosForType,
-  setDemoBorrowSymbol,
-  setDemoEnabled,
-  setDemoHideReal,
-  setDemoItemAmount,
-  setDemoItemBatched,
-  setDemoItemState,
-  setDemoItemType,
-  useDemoBorrowSymbol,
-  useDemoEnabled,
-  useDemoHideReal,
-  useDemoItems,
-} from "./demoDeposit";
-import {
-  PANEL_BUTTON_CLASS,
-  PANEL_SECTION_CLASS,
-  PANEL_SECTION_TITLE_CLASS,
-  panelSegmentClass,
-} from "./panelChrome";
-
-const TYPE_LABELS: Record<DemoType, string> = {
-  deposit: "Deposit",
-  collateral: "Collateral",
-  loan: "Loan",
-  activity: "Activity",
-};
-
-const CTA_BADGE: Record<DemoCta, { label: string; className: string }> = {
-  primary: { label: "Orange CTA", className: "bg-orange-500 text-white" },
-  outlined: {
-    label: "Outlined CTA",
-    className: "border border-zinc-400 text-zinc-200",
-  },
-  none: { label: "No CTA", className: "bg-zinc-700 text-zinc-300" },
-};
-
-/**
- * Deposit "mode" segments over the flat DEPOSIT_SCENARIOS list. The mode select
- * picks a segment; the slider/dropdown scrub within it. Different wallet is a
- * single state (count 1), so its slider is inert.
- */
-const DEPOSIT_SEGMENTS: {
-  mode: string;
-  label: string;
-  offset: number;
-  count: number;
-}[] = [
-  {
-    mode: "normal",
-    label: "Normal",
-    offset: 0,
-    count: DEPOSIT_FLOW_SCENARIO_COUNT,
-  },
-  {
-    mode: "expired",
-    label: "Expired",
-    offset: DEPOSIT_FLOW_SCENARIO_COUNT,
-    count: DEPOSIT_EXPIRED_SCENARIO_COUNT,
-  },
-  {
-    mode: "different-wallet",
-    label: "Different wallet",
-    offset: DEPOSIT_FLOW_SCENARIO_COUNT + DEPOSIT_EXPIRED_SCENARIO_COUNT,
-    count: DEPOSIT_DIFFERENT_WALLET_SCENARIO_COUNT,
-  },
-  {
-    mode: "activation-window",
-    label: "Activation window",
-    offset:
-      DEPOSIT_FLOW_SCENARIO_COUNT +
-      DEPOSIT_EXPIRED_SCENARIO_COUNT +
-      DEPOSIT_DIFFERENT_WALLET_SCENARIO_COUNT,
-    count: DEPOSIT_ACTIVATION_WINDOW_SCENARIO_COUNT,
-  },
-];
-
-function ItemRow({ item, index }: { item: DemoItem; index: number }) {
-  const borrowSymbol = useDemoBorrowSymbol();
-  const scenarios = scenariosForType(item.type, borrowSymbol);
-  const total = scenarios.length;
-  // Deposits pick a "mode" (Normal / Expired / Different wallet); the slider and
-  // dropdown then scrub within that mode's segment. Other types are one flat
-  // segment with no mode select. `stateIndex` still addresses the flat list.
-  const isDeposit = item.type === "deposit";
-  const segment = isDeposit
-    ? (DEPOSIT_SEGMENTS.find(
-        (s) =>
-          item.stateIndex >= s.offset && item.stateIndex < s.offset + s.count,
-      ) ?? DEPOSIT_SEGMENTS[0])
-    : null;
-  const offset = segment?.offset ?? 0;
-  const count = segment?.count ?? total;
-  const localIndex = item.stateIndex - offset;
-  const scenario = scenarios[item.stateIndex] ?? scenarios[0];
-  const badge = CTA_BADGE[scenario.expectedCta];
-  const position = index + 1;
-  const clampLocal = (next: number) => Math.min(count - 1, Math.max(0, next));
-
-  return (
-    <div className={`space-y-2 ${PANEL_SECTION_CLASS}`}>
-      <div className="flex items-center justify-between gap-2">
-        <select
-          value={item.type}
-          onChange={(e) =>
-            setDemoItemType(item.key, e.target.value as DemoType)
-          }
-          className="rounded border border-zinc-600 bg-zinc-800 px-2 py-1 text-xs"
-          aria-label={`Mock ${position} type`}
-        >
-          {(Object.keys(TYPE_LABELS) as DemoType[]).map((type) => (
-            <option key={type} value={type}>
-              {TYPE_LABELS[type]}
-            </option>
-          ))}
-        </select>
-        <div className="flex shrink-0 items-center gap-2">
-          <span className="text-xs tabular-nums text-zinc-400">
-            {localIndex + 1}/{count}
-          </span>
-          <span
-            className={`rounded-full px-2 py-0.5 text-xs font-medium ${badge.className}`}
-          >
-            {badge.label}
-          </span>
-          <button
-            type="button"
-            onClick={() => removeDemoItem(item.key)}
-            className={PANEL_BUTTON_CLASS}
-            aria-label={`Remove mock ${position}`}
-          >
-            ✕
-          </button>
-        </div>
-      </div>
-
-      {/* Mode select (deposit only) sits above the slider: it picks the segment
-          the slider scrubs — Normal (flow), Expired, or Different wallet. */}
-      {isDeposit && segment && (
-        <label className="flex items-center justify-between gap-2 text-xs">
-          <span className="shrink-0">Mode</span>
-          <select
-            value={segment.mode}
-            onChange={(e) => {
-              const next = DEPOSIT_SEGMENTS.find(
-                (s) => s.mode === e.target.value,
-              );
-              if (next) setDemoItemState(item.key, next.offset);
-            }}
-            className="min-w-0 flex-1 rounded border border-zinc-600 bg-zinc-800 px-2 py-1 text-xs"
-            aria-label={`Mock ${position} mode`}
-          >
-            {DEPOSIT_SEGMENTS.map((s) => (
-              <option key={s.mode} value={s.mode}>
-                {s.label}
-              </option>
-            ))}
-          </select>
-        </label>
-      )}
-
-      {/* Slider + dropdown both scrub within the current mode's segment —
-          slider for quick stepping, dropdown for jumping straight to a state. */}
-      <div className="flex items-center gap-2">
-        <button
-          type="button"
-          onClick={() =>
-            setDemoItemState(item.key, offset + clampLocal(localIndex - 1))
-          }
-          disabled={localIndex === 0}
-          className={PANEL_BUTTON_CLASS}
-        >
-          Prev
-        </button>
-        <input
-          type="range"
-          min={0}
-          max={count - 1}
-          value={localIndex}
-          disabled={count <= 1}
-          onChange={(e) =>
-            setDemoItemState(item.key, offset + Number(e.target.value))
-          }
-          className="min-w-0 flex-1 accent-orange-500 disabled:opacity-40"
-          aria-label={`Mock ${position} step`}
-        />
-        <button
-          type="button"
-          onClick={() =>
-            setDemoItemState(item.key, offset + clampLocal(localIndex + 1))
-          }
-          disabled={localIndex === count - 1}
-          className={PANEL_BUTTON_CLASS}
-        >
-          Next
-        </button>
-      </div>
-
-      {/* Current-state readout. The mode select + slider drive the state, so
-          this is just a label — no redundant jump-to-state dropdown. */}
-      <div
-        className="truncate text-xs text-zinc-400"
-        title={scenario.label}
-        aria-label={`Mock ${position} state`}
-      >
-        {scenario.label}
-      </div>
-
-      <label className="flex items-center justify-between gap-2 text-xs">
-        <span>Amount ({amountUnitFor(item, borrowSymbol)})</span>
-        <input
-          type="number"
-          min="0"
-          step="0.0001"
-          value={item.amount}
-          onChange={(e) => setDemoItemAmount(item.key, e.target.value)}
-          className="w-28 rounded border border-zinc-600 bg-zinc-800 px-2 py-1 text-xs"
-          aria-label={`Mock ${position} amount (${amountUnitFor(item, borrowSymbol)})`}
-        />
-      </label>
-
-      {item.type === "deposit" && (
-        <label className="flex cursor-pointer items-center justify-between gap-2 text-xs">
-          <span>Batched (group with other batched deposits)</span>
-          <input
-            type="checkbox"
-            checked={item.batched}
-            onChange={(e) => setDemoItemBatched(item.key, e.target.checked)}
-          />
-        </label>
-      )}
-
-      <div className="text-xs text-zinc-500">
-        Renders in: {itemSectionHint(item)}
-      </div>
-    </div>
-  );
+interface GodModePanelProps {
+  registry: DevPanelTab[];
 }
 
-function DemoControls() {
-  const enabled = useDemoEnabled();
-  const hideReal = useDemoHideReal();
-  const items = useDemoItems();
-  const mockArtifactDownload = useArtifactDownloadMockEnabled();
-  const artifactScenario = useArtifactDownloadScenario();
-  const [clearedReceipts, setClearedReceipts] = useState<number | null>(null);
-  const borrowSymbol = useDemoBorrowSymbol();
-
-  // Publish the resolved deposit-family aggregates for the real dashboard
-  // sections to read. "Inject demo" gates all four: merely adding a mock item
-  // must not inject it until the toggle opts in.
-  useEffect(() => {
-    setDepositOverride(enabled ? buildDepositsDemo(items, hideReal) : null);
-  }, [enabled, items, hideReal]);
-
-  useEffect(() => {
-    setActivityOverride(
-      enabled ? buildActivitiesDemo(items, hideReal, borrowSymbol) : null,
-    );
-  }, [enabled, items, hideReal, borrowSymbol]);
-
-  useEffect(() => {
-    setCollateralOverride(
-      enabled ? buildCollateralsDemo(items, hideReal) : null,
-    );
-  }, [enabled, items, hideReal]);
-
-  useEffect(() => {
-    setLoanOverride(
-      enabled ? buildLoansDemo(items, hideReal, borrowSymbol) : null,
-    );
-  }, [enabled, items, hideReal, borrowSymbol]);
-
-  // Independent of "Inject demo" — see the toggle's own label below.
-  useEffect(() => {
-    setArtifactDownloadOverride(
-      mockArtifactDownload ? demoFetchAndDownloadArtifacts : null,
-    );
-  }, [mockArtifactDownload]);
-
-  return (
-    <div className="space-y-3">
-      <label className="flex cursor-pointer items-center justify-between gap-2 text-sm">
-        <span>Inject demo</span>
-        <input
-          type="checkbox"
-          checked={enabled}
-          onChange={(e) => setDemoEnabled(e.target.checked)}
-        />
-      </label>
-
-      {/* Independent of "Inject demo": the simulated fetch also applies to
-          real vault rows, and demo rows only become downloadable through it. */}
-      <label className="flex cursor-pointer items-center justify-between gap-2 text-sm">
-        <span>Mock artifact download</span>
-        <input
-          type="checkbox"
-          checked={mockArtifactDownload}
-          onChange={(e) => setArtifactDownloadMockEnabled(e.target.checked)}
-        />
-      </label>
-
-      {/* The mock drives the real validator and file sink, so these pick what
-          a hostile or broken vault provider sends back. */}
-      <label
-        className={`flex items-center justify-between gap-2 text-sm ${
-          mockArtifactDownload ? "" : "opacity-40"
-        }`}
-      >
-        <span>Artifact response</span>
-        <select
-          disabled={!mockArtifactDownload}
-          value={artifactScenario}
-          onChange={(e) =>
-            setArtifactDownloadScenario(e.target.value as DemoArtifactScenario)
-          }
-          className="rounded bg-neutral-800 px-2 py-1 text-xs"
-        >
-          {DEMO_ARTIFACT_SCENARIOS.map((scenario) => (
-            <option key={scenario} value={scenario}>
-              {DEMO_ARTIFACT_SCENARIO_LABELS[scenario]}
-            </option>
-          ))}
-        </select>
-      </label>
-
-      {/* A satisfied gate hides the card's download button, so re-testing the
-          flow against the same vault needs the receipts gone. */}
-      <button
-        type="button"
-        onClick={() => {
-          const cleared = clearArtifactDownloadReceipts();
-          setClearedReceipts(cleared);
-        }}
-        className="w-full rounded bg-neutral-800 px-2 py-1 text-left text-sm hover:bg-neutral-700"
-      >
-        {clearedReceipts === null
-          ? "Clear artifact receipts"
-          : `Cleared ${clearedReceipts} artifact receipt(s) — reload`}
-      </button>
-
-      <fieldset
-        disabled={!enabled}
-        className={`space-y-3 border-0 p-0 ${enabled ? "" : "opacity-40"}`}
-      >
-        <label className="flex cursor-pointer items-center justify-between gap-2 text-sm">
-          <span>Hide real items</span>
-          <input
-            type="checkbox"
-            checked={hideReal}
-            onChange={(e) => setDemoHideReal(e.target.checked)}
-          />
-        </label>
-
-        {/* Drives both the loan mock's row and the debt-denominated activity
-            mocks (Borrow / Repay / liquidation), so a loan and its matching
-            activity row never disagree about the asset. */}
-        <label className="flex items-center justify-between gap-2 text-sm">
-          <span>Borrowed asset (loan + activity rows)</span>
-          <select
-            value={borrowSymbol}
-            onChange={(e) =>
-              setDemoBorrowSymbol(e.target.value as DemoBorrowSymbol)
-            }
-            className="rounded border border-zinc-600 bg-zinc-800 px-2 py-1 text-xs"
-            aria-label="Borrowed asset"
-          >
-            {DEMO_BORROW_SYMBOL_OPTIONS.map((symbol) => (
-              <option key={symbol} value={symbol}>
-                {symbol}
-              </option>
-            ))}
-          </select>
-        </label>
-
-        {items.map((item, index) => (
-          <ItemRow key={item.key} item={item} index={index} />
-        ))}
-
-        <button
-          type="button"
-          onClick={() => addDemoItem("deposit")}
-          className="w-full rounded border border-dashed border-zinc-600 px-2 py-1.5 text-xs text-zinc-300 hover:bg-zinc-800"
-        >
-          + Add mock
-        </button>
-      </fieldset>
-    </div>
-  );
-}
-
-/** Dev theme switch — drives the app's next-themes provider so the dashboard
- *  behind the panel can be previewed in light / dark / system. */
-const THEME_OPTIONS = ["light", "dark", "system"] as const;
-
-function ThemeControls() {
-  const { theme, setTheme } = useTheme();
-  return (
-    <div className="space-y-2">
-      <div className={PANEL_SECTION_TITLE_CLASS}>Theme</div>
-      <div className="flex gap-2">
-        {THEME_OPTIONS.map((option) => (
-          <button
-            key={option}
-            type="button"
-            onClick={() => setTheme(option)}
-            className={`${panelSegmentClass(theme === option)} capitalize`}
-          >
-            {option}
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-/**
- * The three notifications that are NOT derived from the liquidation-cascade
- * calculation (Figma v3 §7 soft-paused / §8 fully paused / §9 maximum vaults
- * reached). They are driven by live chain reads — `useVaultCountCap` and
- * `useProtocolGateState` — so each gets a store override here that the
- * rendering component prefers over its live value.
- *
- * These live in the always-mounted panel body rather than inside the
- * position-notifications section, because they are independent of the
- * liquidation-notifications flag that gates that section. The overrides survive
- * collapsing the panel on purpose: you collapse it to look at the card.
- */
-const PROTOCOL_STATUS_LABELS: Record<ProtocolStatus, string> = {
-  frozen: "Soft-paused",
-  paused: "Fully paused",
-};
-
-function SegmentButton({
-  label,
-  active,
-  onClick,
+function PanelHeader({
+  visibleTabs,
+  activeTabId,
+  onSelectTab,
 }: {
-  label: string;
-  active: boolean;
-  onClick: () => void;
+  visibleTabs: DevPanelTab[];
+  activeTabId: string;
+  onSelectTab: (id: string) => void;
 }) {
   return (
-    <button
-      type="button"
-      onClick={onClick}
-      className={panelSegmentClass(active)}
-    >
-      {label}
-    </button>
-  );
-}
-
-function NotificationOverrideControls() {
-  const maxVaultsOverride = useDebugMaxVaultsOverride();
-  const protocolStatusOverride = useDebugProtocolStatusOverride();
-
-  useEffect(() => {
-    setMaxVaultsOverride(maxVaultsOverride);
-  }, [maxVaultsOverride]);
-
-  useEffect(() => {
-    setProtocolStatusOverride(protocolStatusOverride);
-  }, [protocolStatusOverride]);
-
-  return (
     <div className="space-y-2">
-      <div className={PANEL_SECTION_TITLE_CLASS}>Notifications</div>
-
-      <label className="flex cursor-pointer items-center justify-between gap-2 text-sm">
-        <span>Maximum vaults reached</span>
-        <input
-          type="checkbox"
-          checked={maxVaultsOverride !== null}
-          onChange={(e) =>
-            setDebugMaxVaultsOverride(
-              e.target.checked ? DEBUG_FORCED_MAX_VAULTS : null,
-            )
-          }
-        />
-      </label>
-
-      <div className="space-y-1">
-        <div className="text-xs text-zinc-400">Protocol status</div>
-        <div className="flex gap-2">
+      <SummaryChip />
+      <div className="flex flex-wrap gap-1.5">
+        {visibleTabs.map((tab) => (
           <SegmentButton
-            label="Live"
-            active={protocolStatusOverride === null}
-            onClick={() => setDebugProtocolStatusOverride(null)}
+            key={tab.id}
+            label={tab.label}
+            active={tab.id === activeTabId}
+            onClick={() => onSelectTab(tab.id)}
           />
-          {(Object.keys(PROTOCOL_STATUS_LABELS) as ProtocolStatus[]).map(
-            (status) => (
-              <SegmentButton
-                key={status}
-                label={PROTOCOL_STATUS_LABELS[status]}
-                active={protocolStatusOverride === status}
-                onClick={() => setDebugProtocolStatusOverride(status)}
-              />
-            ),
-          )}
-        </div>
+        ))}
       </div>
     </div>
   );
 }
 
-/**
- * Loans-summary overrides (v3 /loans). The row list is driven by "Loan" mocks
- * in the Mocks section; these two cover the summary states a real position
- * can't be talked into on demand — a health factor in each band, and the
- * borrow-capacity cards while their read is loading or has failed.
- */
-const BORROW_CAPACITY_LABELS: Record<DebugBorrowCapacityState, string> = {
-  loading: "Loading",
-  error: "Error",
-};
-
-function LoansOverrideControls() {
-  const healthFactorOverride = useDebugHealthFactorOverride();
-  const borrowCapacityOverride = useDebugBorrowCapacityStateOverride();
-  // Resolved {loading, error} | null snapshot — see debugPositionStore's
-  // DEBUG_BORROW_CAPACITY_SNAPSHOTS for why the Error stays dev-only there.
-  const resolvedBorrowCapacity = useDebugBorrowCapacity();
-
-  useEffect(() => {
-    setHealthFactorOverride(healthFactorOverride);
-  }, [healthFactorOverride]);
-
-  useEffect(() => {
-    setBorrowCapacityOverride(resolvedBorrowCapacity);
-  }, [resolvedBorrowCapacity]);
-
-  return (
-    <div className="space-y-2">
-      <div className={PANEL_SECTION_TITLE_CLASS}>Loans summary</div>
-
-      <div className="space-y-1">
-        <div className="text-xs text-zinc-400">Health factor</div>
-        <div className="flex gap-2">
-          <SegmentButton
-            label="Live"
-            active={healthFactorOverride === null}
-            onClick={() => setDebugHealthFactorOverride(null)}
-          />
-          {DEBUG_HEALTH_FACTORS.map(({ value, label }) => (
-            <SegmentButton
-              key={label}
-              label={`${label} (${value})`}
-              active={healthFactorOverride === value}
-              onClick={() => setDebugHealthFactorOverride(value)}
-            />
-          ))}
-        </div>
-      </div>
-
-      <div className="space-y-1">
-        <div className="text-xs text-zinc-400">Borrow capacity</div>
-        <div className="flex gap-2">
-          <SegmentButton
-            label="Live"
-            active={borrowCapacityOverride === null}
-            onClick={() => setDebugBorrowCapacityStateOverride(null)}
-          />
-          {(
-            Object.keys(BORROW_CAPACITY_LABELS) as DebugBorrowCapacityState[]
-          ).map((state) => (
-            <SegmentButton
-              key={state}
-              label={BORROW_CAPACITY_LABELS[state]}
-              active={borrowCapacityOverride === state}
-              onClick={() => setDebugBorrowCapacityStateOverride(state)}
-            />
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function PanelBody({ children }: { children?: ReactNode }) {
+function PanelBody({
+  visibleTabs,
+  activeTabId,
+}: {
+  visibleTabs: DevPanelTab[];
+  activeTabId: string;
+}) {
   return (
     <div className="space-y-4">
-      <ThemeControls />
-      <NotificationOverrideControls />
-      <LoansOverrideControls />
-      <div className="space-y-2">
-        <div className={PANEL_SECTION_TITLE_CLASS}>Mocks</div>
-        <DemoControls />
-      </div>
-      {children}
+      {visibleTabs.map((tab) => (
+        <div key={tab.id} hidden={tab.id !== activeTabId}>
+          <tab.Component />
+        </div>
+      ))}
+      <CascadeOverridePublisher />
     </div>
   );
 }
 
-export function GodModePanel({ children }: { children?: ReactNode }) {
+export function GodModePanel({ registry }: GodModePanelProps) {
+  const visibleTabs = registry.filter((tab) => !tab.gate || tab.gate());
+  const [activeTabId, setActiveTabId] = useState(
+    () => visibleTabs[0]?.id ?? "",
+  );
+  const activeTab = visibleTabs.some((tab) => tab.id === activeTabId)
+    ? activeTabId
+    : (visibleTabs[0]?.id ?? "");
+
   // Defaults: collapsed (small launcher) anchored bottom-right. `pos` is null
   // until the user drags — then it switches to absolute top/left positioning.
   const [collapsed, setCollapsed] = useState(true);
@@ -771,7 +186,14 @@ export function GodModePanel({ children }: { children?: ReactNode }) {
             Return ↙
           </button>
         </div>
-        <PanelBody>{children}</PanelBody>
+        <div className="mb-3">
+          <PanelHeader
+            visibleTabs={visibleTabs}
+            activeTabId={activeTab}
+            onSelectTab={setActiveTabId}
+          />
+        </div>
+        <PanelBody visibleTabs={visibleTabs} activeTabId={activeTab} />
       </div>,
       popupContainer,
     );
@@ -829,8 +251,16 @@ export function GodModePanel({ children }: { children?: ReactNode }) {
           </button>
         </div>
 
+        <div className="shrink-0 border-b border-zinc-800 p-3">
+          <PanelHeader
+            visibleTabs={visibleTabs}
+            activeTabId={activeTab}
+            onSelectTab={setActiveTabId}
+          />
+        </div>
+
         <div className="min-h-0 flex-1 overflow-y-auto overflow-x-hidden p-4">
-          <PanelBody>{children}</PanelBody>
+          <PanelBody visibleTabs={visibleTabs} activeTabId={activeTab} />
         </div>
       </div>
     );
